feat: MinMax widget
